requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
pytest
httpx
pytest-asyncio
passlib[bcrypt]
python-jose[cryptography]
python-multipart
beautifulsoup4
sqlalchemy>=2.0.0
asyncpg
